Add buggy grenade prediction
Currently hooks right after the PIP prediction, which means it is not
able to reset the drawn prediction. It draws the original prediction,
but the code for removing the original is already there. The code is a
bit ugly and uses a bunch of byte pattern signatures which are likely to
break since the few versions I have databases for all have different
ones.