Add CRC self check and Emergency Crash system
Proofing that your code has purpose may seem difficult sometimes, but on
others it proofs its usefulness by itself. This self check is certainly
one of these systems because I got quite scared when I realized that the
CRC does not update at all.

This was caused by getting the pointer from the move_crc when the
function I was passing it, was trying to get the pointer again.
Basically resulting in writing the CRC one level of indirection too
deep.

I then realized that the CRC is actually wrong, it was too short. I
found a website that actually decoded protobuf messages directly on
which I noticed that my CRCs are missing the viewangle components which
are zero. I fixed that as well and now the CRCs seem to match.

Despite fixing those issues, I don't know if I can somehow proof on the
self check that the CRC is actually regenerated. I noticed that it
wasn't by checking if it changed after the actual recalculation. Perhaps
I can check if the UserCmd changed and if it did then force the CRC to
also change, however I'm still quite unsure how safe it is to clone a
protobuf object. Anyways here's the website for whoever may be reading
this: https://www.terrific.tools/data/protobuf-decoder

Note, that I'm not associated with that website, nor do I endorse it,
however it helped me.