Add crude subtick auto strafe implementation
This adds a crude subtick strafe implementation.

There are several things wrong with this, for one I don't see a reason
for why I need to reset the forward and left move from the basecmd.

I also discovered that the calculation for the optimal angle is far from
correct. I took it from FrameworkCSGO, which in turn took it from NEPS
as suggested by the comment next to it, however I never actually did the
math myself, so this seems to come back and bite me now. The subtick
strafer is thus not as fast as it could be.

I also yolo'd the velocity prediction essentially, it works for now but
is obviously horribly wrong. I will emulate AirMove properly once I feel
like updating this.

On a positive note, I no longer initialize the convars ad-hoc, so the
first jump no longer causes a micro stutter, in which the convars need
to be looked up. I also cleaned up the code a lot, because it was just
all over the place before.

Anyways, this is still a WIP, just like denoted by the settings name.
Eventually I would like to add subtick support to the basic forward
strafer as well, although I believe the intersection of people using the
more legit strafer and the less legit subtick abuse would be rather
minimal.

Anyways I could use some help on basically all the points here, but I'm
just happy that I have anything that works right now, because it took
quite a while for me to wrap my head around this subtick moves system.