Add some defensive programming to prevent the model changer from crashing uncontrollably
Previously the model changer would crash when force updating, this seems
to solve this and should in general make the entire thing much more
resiliant to initial conditions.

I also added a check that checks if the model is already the one it
wants to change it to, in that case I just don't run the model changer
at all, because there is no point in doing so. This should also save
some frame time