Revamp the entire model selection with fancy model pictures
This took so much time and sanity.

The VTex decoder barely supports any formats, but seemingly the one I
need for those images.

When reading the vtex file from the vpk, I actually clone it inside a
vector to get rid of the valve data structures, but this is obviously
imperfect. I would prefer if the api inside the VTexDecoder was a little
saner, but it works for now.

I also no longer need the agent name so the save/load logic is much
better.

I'm still not sure how to deal with untranslated models, if you have
ideas, I'd be happy to hear them.