Compress raw image data
I recently thought that if we already have LZ4 to decompress valve's
images, perhaps we could also re-use it to compress our own images.

I implemented it and most images seem to get reduced to only 30% of the
size of their former selves. The decompression is also fairly fast, so
you don't even notice it when loading the texture on the GPU.

Pretty cool LZ4, pretty cool.