Prevent Auto pickup from grabbing guns through walls
The game is ignorant to the fact that the pickup function used in the
server code checks for a LOS, this causes the bug that sometimes guns
cant be picked up in the menu. This commit addresses this by
implementing the same check into the Auto Pickup albeit a bit more hacky
than what the game is probably doing.