Remove NodeCircuits and replace them with normal keybinds
I might add the reasoning for this removal here.
TL;DR I tried to solve a problem, that is already solved and my solution
was flawed.

Long version:
There is a lot of software that advertises some feature that is deeply
flawed in some way, for example subticks to stay on topic with CS2.
I was trying to make NodeCircuits allow anyone to hook into the features
and change their behavior. This is already a solved problem since all of
this is open source. Now the only remaining reasoning is simplicity and
ease of use. Neither of which would have worked out since the node
circuits don't respect thread-safety, which makes them flawed by
themselves. And even if they would then they are certainly not simple.
I'd imagine that it would amount to me reimplementing Scratch for CS2,
which I don't think is a good approach to this. So to provide a good
piece of software I think I need to overcome priding myself in a feature
that I know wouldn't work out in the longrun and remove it.